2.4 - Code Implemented - 9

+++ Programming Contest : Data Structure & Algorithm by Md. Mahbubul Hasan/Chapter 02/2.4/Exercise9.c
@@ -9,7 +9,7 @@ int main()
     while (currentPalindromeNo != n)
     {
         long long workingWith = currentlyConsidering;
-        int reversed = 0;
+        long long reversed = 0;
         while (workingWith != 0)
         {
             reversed = reversed * 10;
